requirements.txt
feedparser
tweepy
python-dotenv
openai
sumy
setuptools
nltk
